requirements.txt
# Tushare金融数据API
tushare>=1.3.0

# 飞书API
requests>=2.28.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 技术指标计算
pandas-ta>=0.3.14b0

# 配置文件解析
pyyaml>=6.0

# 周进度条显示
tqdm>=4.65.0

# 开发依赖
pytest>=7.0.0
pytest-cov>=4.0.0
